+++ Sources/XMTPiOS/Extensions/Date.swift
@@ -11,4 +11,8 @@ extension Date {
 	var millisecondsSinceEpoch: Double {
 		timeIntervalSince1970 * 1000
 	}
+
+	init(millisecondsSinceEpoch: Int64) {
+		self.init(timeIntervalSince1970: TimeInterval(millisecondsSinceEpoch / 1_000_000_000))
+	}
 }
